drop pinia store for words, just use everything directly
