+++ src/App.vue
@@ -7,14 +7,3 @@ import { RouterView } from 'vue-router'
   <Nav />
   <router-view></router-view>
 </template>
-
-<script>
-import { useWordsStore } from './stores/words'
-
-export default {
-  async mounted() {
-    await useWordsStore().populate()
-  }
-}
-
-</script>
