+++ src/main.js
@@ -1,12 +1,9 @@
 import './assets/css/main.css'
 import { createApp } from 'vue'
-import { createPinia } from 'pinia'
 import router from './router'
 import App from './App.vue'
 
 const app = createApp(App)
-const pinia = createPinia()
 
-app.use(pinia)
 app.use(router)
 app.mount('#app')
